06/requirements.txt
#ex00
grpcio-tools
grpcio
protobuf
argparse
#ex01
pydantic
#ex02
sqlalchemy
psycopg2
